updated the stepwat executable call
In the latest branch of STEPWAT stand alone (SoilWat31_drs), we know
have the capacity to track and output SOILWAT variables. There is now a
different executable call to do implement this and I made this the
default, so we will always get SOILWAT output when running the
StepWat_R_Wrapper.